final project

+++ TeamMember.cs
@@ -2,5 +2,5 @@
 public class TeamMember {
     public string Name { get; set; }
     public int SkillLevel { get; set; }
-    public decimal CourageFactor { get; set; }
+    public double CourageFactor { get; set; }
 }
